index view added to browse the videos in paginated manner

+++ youtube_api_fetcher/views.py
@@ -1,19 +1,34 @@
 from django.shortcuts import render
-
-# Create your views here.
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.http import HttpResponse
+from background_task.models import Task
 
-from youtube_api_fetcher.cron import fetch_from_youtube
-
+from .cron import fetch_from_youtube
 from .models import Video
 
-from background_task.models import Task
+# clear previous tasks
+Task.objects.all().delete()
 
 def cron(request):
     # clear previous tasks
-    Task.objects.all().delete()
+    if Task.objects.count() > 0:
+        Task.objects.all().delete()
+
     fetch_from_youtube(repeat = 30)
-    return HttpResponse("Cron Job Started")
+    return HttpResponse("<h1>Cron Job Started<h1>")
 
 def index(request):
-    return HttpResponse("Project Started")
+    # Getting all videos in descending order
+    video_list = Video.objects.all().order_by('-publishedAt')
+    page = request.GET.get('page', 1)
+
+    # Pagination with 10 values per page
+    paginator = Paginator(video_list, 10)
+    try:
+        videos = paginator.page(page)
+    except PageNotAnInteger:
+        videos = paginator.page(1)
+    except EmptyPage:
+        videos = paginator.page(paginator.num_pages)
+
+    return render(request, 'index.html', { 'videos': videos })
